+++ functions/New-MySQLiteDBTable.ps1
@@ -3,29 +3,47 @@
 Function New-MySQLiteDBTable {
     [cmdletbinding(SupportsShouldProcess, DefaultParameterSetName = "filetyped")]
     [alias("New-DBTable", "ndbt")]
-    [outputtype("None")]
+    [OutputType("None")]
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the path to the SQLite database file.", ValueFromPipelineByPropertyName, ParameterSetName = "filetyped")]
+        [Parameter(
+            Position = 0,
+            Mandatory,
+            HelpMessage = "Enter the path to the SQLite database file.",
+            ValueFromPipelineByPropertyName,
+            ParameterSetName = "filetyped"
+        )]
         [Parameter(ParameterSetName = "filenamed")]
         [Alias("fullname", "database")]
         [ValidateNotNullOrEmpty()]
         [string]$Path,
 
-        [Parameter(HelpMessage = "Specify an existing open database connection.", ParameterSetName = "cnxtyped")]
+        [Parameter(
+            HelpMessage = "Specify an existing open database connection.",
+            ParameterSetName = "cnxtyped"
+        )]
         [Parameter(ParameterSetName = "cnxnamed")]
         [ValidateNotNullOrEmpty()]
         [System.Data.SQLite.SQLiteConnection]$Connection,
 
-        [Parameter(Mandatory, HelpMessage = "Enter the name of the new table. Table names are technically case-sensitive.")]
+        [Parameter(
+            Mandatory,
+            HelpMessage = "Enter the name of the new table. Table names are technically case-sensitive."
+        )]
         [ValidateNotNullOrEmpty()]
         [string]$TableName,
 
-        [parameter( HelpMessage = "Enter an ordered hashtable of column definitions", ParameterSetName = "filetyped")]
+        [parameter(
+            HelpMessage = "Enter an ordered hashtable of column definitions",
+            ParameterSetName = "filetyped"
+        )]
         [Parameter(ParameterSetName = "cnxtyped")]
         [ValidateNotNullOrEmpty()]
         [System.Collections.Specialized.OrderedDictionary]$ColumnProperties,
 
-        [parameter( HelpMessage = "Enter an array of column names.", ParameterSetName = "cnxnamed")]
+        [parameter(
+            HelpMessage = "Enter an array of column names.",
+            ParameterSetName = "cnxnamed"
+        )]
         [Parameter(ParameterSetName = "filenamed")]
         [ValidateNotNullOrEmpty()]
         [string[]]$ColumnNames,
@@ -36,13 +54,16 @@ Function New-MySQLiteDBTable {
         [Parameter(HelpMessage = "Overwrite an existing table. This could result in data loss.")]
         [switch]$Force,
 
-        [Parameter(HelpMessage = "Keep an existing connection open.", ParameterSetName = "cnxtyped")]
+        [Parameter(
+            HelpMessage = "Keep an existing connection open.",
+            ParameterSetName = "cnxtyped"
+        )]
         [Parameter(ParameterSetName = "cnxnamed")]
         [switch]$KeepAlive
     )
 
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] $($MyInvocation.MyCommand)"
     } #begin
     Process {
         if ($Path) {
@@ -60,22 +81,22 @@ Function New-MySQLiteDBTable {
         else {
             Write-Verbose "[$((Get-Date).TimeOfDay)] Using an existing connection"
         }
-        if ($connection.state -eq 'Open' -OR $PSBoundparameters.ContainsKey("WhatIf")) {
+        if ($connection.state -eq 'Open' -OR $PSBoundParameters.ContainsKey("WhatIf")) {
 
             $cmd = $connection.CreateCommand()
             #test if table already exists
             if (-not $Force) {
-                $cmd.CommandText = "SELECT name FROM sqlite_master WHERE type='table' AND name='$Tablename' COLLATE NOCASE;"
-                Write-Verbose "[$((Get-Date).TimeOfDay)] Using table $Tablename"
+                $cmd.CommandText = "SELECT name FROM sqlite_master WHERE type='table' AND name='$TableName' COLLATE NOCASE;"
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Using table $TableName"
                 Write-Verbose "[$((Get-Date).TimeOfDay)] $($cmd.CommandText)"
 
                 $ds = New-Object System.Data.DataSet
                 $da = New-Object System.Data.SQLite.SQLiteDataAdapter($cmd)
                 [void]$da.fill($ds)
 
-                if ($ds.Tables.rows.name -contains $Tablename) {
-                    Write-Warning "The table $Tablename already exists. Use -Force to overwrite it."
-                    $tableFree = $False
+                if ($ds.Tables.rows.name -contains $TableName) {
+                    Write-Warning "The table $TableName already exists. Use -Force to overwrite it."
+                    $TableFree = $False
                 }
                 else {
                     $TableFree = $True
@@ -83,23 +104,23 @@ Function New-MySQLiteDBTable {
             } #if -not -Force
             else {
                 #drop the table
-                $query = "DROP TABLE IF EXISTS $Tablename;"
+                $query = "DROP TABLE IF EXISTS $TableName;"
                 $cmd.CommandText = $query
-                if ($pscmdlet.ShouldProcess($query)) {
+                if ($PSCmdlet.ShouldProcess($query)) {
                     [void]$cmd.ExecuteNonQuery()
                 }
-                $tablefree = $True
+                $TableFree = $True
             }
 
             If ($TableFree ) {
-                Write-Verbose "[$((Get-Date).TimeOfDay)] Creating table $Tablename"
-                [string]$query = "CREATE TABLE $tablename "
+                Write-Verbose "[$((Get-Date).TimeOfDay)] Creating table $TableName"
+                [string]$query = "CREATE TABLE $TableName "
 
-                if ($pscmdlet.ParameterSetName -match 'typed') {
+                if ($PSCmdlet.ParameterSetName -match 'typed') {
                     $keys = $ColumnProperties.Keys
                     #9/9/2022 Need to let the user specify the primary key
                     #property. Issue #13 JDH
-                    if (-Not ($PSBoundparameters.ContainsKey("Primary"))) {
+                    if (-Not ($PSBoundParameters.ContainsKey("Primary"))) {
                         $primary = $keys | Select-Object -First 1`
                         $cols = $keys | Select-Object -Skip 1
 
@@ -127,7 +148,7 @@ Function New-MySQLiteDBTable {
                 $cmd = $connection.CreateCommand()
                 $cmd.CommandText = $query
                 Write-Verbose "[$((Get-Date).TimeOfDay)] $query"
-                if ($pscmdlet.ShouldProcess($query)) {
+                if ($PSCmdlet.ShouldProcess($query)) {
                     [void]$cmd.ExecuteNonQuery()
                 }
             }
@@ -141,6 +162,6 @@ Function New-MySQLiteDBTable {
             Write-Verbose "[$((Get-Date).TimeOfDay)] Closing database connection"
             closedb -connection $connection -cmd $cmd
         }
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($MyInvocation.MyCommand)"
     } #end
 }