+++ functions/private.ps1
@@ -6,8 +6,8 @@ Function resolvedb {
 
     Write-Verbose "[$((Get-Date).TimeOfDay)] ResolveDB Resolving $path"
     #resolve or convert path into a full filesystem path
-    $path = $executioncontext.sessionstate.path.GetUnresolvedProviderPathFromPSPath($path)
-    [pscustomobject]@{
+    $path = $ExecutionContext.SessionState.path.GetUnresolvedProviderPathFromPSPath($path)
+    [PSCustomObject]@{
         Path   = $path
         Exists = Test-Path -Path $path
     }
@@ -40,23 +40,24 @@ Function closedb {
         $connection.Dispose()
     }
 }
+
 Function buildquery {
     [cmdletbinding()]
     Param(
         [parameter(Mandatory)]
         [object]$InputObject,
         [parameter(Mandatory)]
-        [string]$Tablename
+        [string]$TableName
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($MyInvocation.MyCommand)"
     } #begin
 
     Process {
         #9/9/2022 Need to insert property names with a dash in []
         #this should fix Issue #14 JDH
         $list = [System.Collections.Generic.list[string]]::new()
-        foreach ($n in $InputObject.psobject.properties.name) {
+        foreach ($n in $InputObject.PSObject.properties.name) {
             if ($n -match "^\S+\-\S+$") {
              #   write-host "REPLACE DASHED $n" -ForegroundColor RED
                 $n =   "[{0}]" -f $matches[0]
@@ -65,17 +66,147 @@ Function buildquery {
             $list.add($n)
         }
         $names = $list -join ","
-        #$names = $InputObject.psobject.Properties.name -join ","
+        #$names = $InputObject.PSObject.Properties.name -join ","
 
-        $inputobject.psobject.Properties | ForEach-Object -Begin {
+        $InputObject.PSObject.Properties | ForEach-Object -Begin {
             $arr = [System.Collections.Generic.list[string]]::new()
         } -Process {
-            if ($_.TypeNameofValue -match "String|Int\d{2}|Double|Datetime|Long") {
+            if ($_.TypeNameOfValue -match "String|Int\d{2}|Double|DateTime|Long") {
                 #9/12/2022 need to escape values that might have single quote
                 $v = $_.Value -replace "'","''"
                 $arr.Add(@(, $v))
             }
-            elseif ($_.TypeNameofValue -match "Boolean") {
+            elseif ($_.TypeNameOfValue -match "Boolean") {
+                #turn Boolean into an INT
+                $arr.Add(@(, ($_.value -as [int])))
+            }
+            else {
+                #only create an entry if there is a value
+                if ($null -ne $_.value) {
+                    Write-Verbose "[$((Get-Date).TimeOfDay)] Creating cliXML for a blob"
+                    $in = ($_.value | ConvertTo-CliXml) -replace "'","''"
+                    $arr.Add(@(, "$($in)"))
+                }
+                else {
+                    $arr.Add("")
+                }
+            }
+        }
+        $values = $arr -join "','"
+      #   If ($names.split(".").count -eq ($values -split "','").count) {
+             "Insert Into $TableName ($names) values ('$values')"
+             #$global:q= "Insert Into $TableName ($names) values ('$values')"
+             #$global:n = $names
+             #$global:v = $values
+       #  }
+        # else {
+        #    Write-Warning "There is a mismatch between the number of column headings ($($names.split(".").count)) and values ($(($values -split "','").count))"
+        # }
+    } #process
+
+    End {
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($MyInvocation.MyCommand)"
+
+    } #end
+
+} #close buildquery
+
+
+Function frombytes {
+    [cmdletbinding()]
+    Param([byte[]]$Bytes)
+
+    #only process if there are bytes
+    # Issue #3 7/20/2022 JDH
+    if ($bytes.count -gt 0) {
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Converting from bytes to object"
+        $tmpFile = [system.io.path]::GetTempFileName()
+        [text.encoding]::UTF8.GetString($bytes) | Out-File -FilePath $tmpfile -Encoding utf8
+        Import-Clixml -Path $tmpFile
+        if (Test-Path $tmpfile) {
+            Remove-Item $tmpFile
+        }
+    }
+}
+
+# cliXML functions added 20 Feb 2023 from Issue #16
+# functions authored by https://github.com/SeSeKenny
+function ConvertTo-CliXml {
+    [CmdletBinding()]
+    param (
+        [Parameter(ValueFromPipeline=$true)]
+        $Object
+    )
+    begin {
+        $Objects=@()
+    }
+    process {
+        $Objects+=$Object
+    }
+    end {
+        if ($Objects.Count -eq 1) {$Objects=$Objects[0]}
+        [System.Management.Automation.PSSerializer]::Serialize($Objects)
+    }
+}
+
+function ConvertFrom-CliXml {
+    [CmdletBinding()]
+    param (
+        [Parameter(ValueFromPipeline)]
+        $Object
+    )
+    begin {
+        $Objects=@()
+    }
+    process {
+        $Objects+=$Object
+    }
+    end {
+        [System.Management.Automation.PSSerializer]::Deserialize($Objects)
+    }
+}
+
+#endregion
+
+<#
+archived
+
+Function OLD-buildquery {
+    [cmdletbinding()]
+    Param(
+        [parameter(Mandatory)]
+        [object]$InputObject,
+        [parameter(Mandatory)]
+        [string]$TableName
+    )
+    Begin {
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($MyInvocation.MyCommand)"
+    } #begin
+
+    Process {
+        #9/9/2022 Need to insert property names with a dash in []
+        #this should fix Issue #14 JDH
+        $list = [System.Collections.Generic.list[string]]::new()
+        foreach ($n in $InputObject.PSObject.properties.name) {
+            if ($n -match '^\S+\-\S+$') {
+                #   write-host "REPLACE DASHED $n" -ForegroundColor RED
+                $n = '[{0}]' -f $matches[0]
+            }
+            # Write-host "ADDING $n" -ForegroundColor CYAN
+            $list.add($n)
+        }
+        $names = $list -join ','
+        #$names = $InputObject.PSObject.Properties.name -join ","
+
+        $InputObject.PSObject.Properties | ForEach-Object -Begin {
+            $arr = [System.Collections.Generic.list[string]]::new()
+        } -Process {
+            if ($_.TypeNameOfValue -match 'String|Int\d{2}|Double|DateTime|Long') {
+                #9/12/2022 need to escape values that might have single quote
+                $v = $_.Value -replace "'", "''"
+                $arr.Add(@(, $v))
+            }
+            elseif ($_.TypeNameOfValue -match 'Boolean') {
                 #turn Boolean into an INT
                 $arr.Add(@(, ($_.value -as [int])))
             }
@@ -90,99 +221,32 @@ Function buildquery {
                     $_.value | Export-Clixml -Path $out -Encoding UTF8 #-Depth 1
                     #for testing
                     # Copy-Item -path $out -Destination d:\temp\out.xml
-                    $in = (Get-Content -Path $out -Encoding UTF8 -ReadCount 0 -Raw) -replace "'","''"
+                    $in = (Get-Content -Path $out -Encoding UTF8 -ReadCount 0 -Raw) -replace "'", "''"
                     $arr.Add(@(, "$($in)"))
                     Remove-Item -Path $out
                 }
                 else {
-                    $arr.Add("")
+                    $arr.Add('')
                 }
             }
         }
         $values = $arr -join "','"
-      #   If ($names.split(".").count -eq ($values -split "','").count) {
-             "Insert Into $Tablename ($names) values ('$values')"
-             #$global:q= "Insert Into $Tablename ($names) values ('$values')"
-             #$global:n = $names
-             #$global:v = $values
-       #  }
+        #   If ($names.split(".").count -eq ($values -split "','").count) {
+        "Insert Into $TableName ($names) values ('$values')"
+        #$global:q= "Insert Into $TableName ($names) values ('$values')"
+        #$global:n = $names
+        #$global:v = $values
+        #  }
         # else {
         #    Write-Warning "There is a mismatch between the number of column headings ($($names.split(".").count)) and values ($(($values -split "','").count))"
         # }
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
+        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($MyInvocation.MyCommand)"
 
     } #end
 
 } #close buildquery
 
-Function OLD-buildquery {
-    [cmdletbinding()]
-    Param(
-        [parameter(Mandatory)]
-        [object]$InputObject,
-        [parameter(Mandatory)]
-        [string]$Tablename
-    )
-    Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Starting $($myinvocation.mycommand)"
-    } #begin
-
-    Process {
-        $names = $InputObject.psobject.Properties.name -join ","
-
-        $inputobject.psobject.Properties | ForEach-Object -Begin { $arr = @() } -Process {
-            if ($_.TypeNameofValue -match "String|Int\d{2}|Double|Datetime|long") {
-                $arr += @(, $_.Value)
-            }
-            elseif ($_.TypeNameofValue -match "Boolean") {
-                #turn Boolean into an INT
-                $arr += @(, ($_.value -as [int]))
-            }
-            else {
-                #only create an entry if there is a value
-                if ($null -ne $_.value) {
-                    Write-Verbose "[$((Get-Date).TimeOfDay)] Creating cliXML for a blob"
-                    #create a temporary cliXML file
-                    $out = [system.io.path]::GetTempFileName()
-                    $_.value | Export-Clixml -Path $out -Encoding UTF8
-                    $in = Get-Content -Path $out -Encoding UTF8 -ReadCount 0 -Raw
-                    $arr += @(, "$($in)")
-                    Remove-Item -Path $out
-                }
-                else {
-                    $arr += ""
-                }
-            }
-        }
-        $values = $arr -join "','"
-
-        "Insert Into $Tablename ($names) values ('$values')"
-
-    } #process
-
-    End {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Ending $($myinvocation.mycommand)"
-
-    } #end
-
-} #close buildquery
-Function frombytes {
-    [cmdletbinding()]
-    Param([byte[]]$Bytes)
-
-    #only process if there are bytes
-    # Issue #3 7/20/2022 JDH
-    if ($bytes.count -gt 0) {
-        Write-Verbose "[$((Get-Date).TimeOfDay)] Converting from bytes to object"
-        $tmpFile = [system.io.path]::GetTempFileName()
-        [text.encoding]::UTF8.getstring($bytes) | Out-File -FilePath $tmpfile -Encoding utf8
-        Import-Clixml -Path $tmpFile
-        if (Test-Path $tmpfile) {
-            Remove-Item $tmpFile
-        }
-    }
-}
-#endregion
+#>